refactor: split component hierarchy editor
